requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
reportlab==4.2.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
